+++ docs/changelog.md
@@ -16,6 +16,7 @@ Click on the version in the following list to see all changes since the previous
 * feat: add warning for invalid user provided as subtask watcher (#64)
 * feat: add support for custom fields of type checkbox (#69)
 * feat: add support for custom fields of type url (#68)
+* feat: add support for custom fields of type user (#66)
 
 ### [23.02.2] - 2023-02-26